+++ Framework/PWMapKit.framework/Versions/A/Headers/PWBuilding.h
@@ -2,19 +2,21 @@
 //  PWBuilding.h
 //  PWMapKit
 //
-//  Copyright (c) 2014 Phunware. All rights reserved.
+//  Copyright (c) 2015 Phunware. All rights reserved.
 //
 
 /**
- The building data encompasses a variety of metadata about the building.
+ A building's data includes a variety of metadata about the building.
  */
 
 #import <CoreLocation/CLLocation.h>
 
 #import "PWMappingTypes.h"
 
+@class PWBuildingFloor;
+
 /**
- The building object encompasses a wide variety of information associated with the building. Information includes but is not limited to floor data, floor resource data, street address and location.
+ The building object encompasses a wide variety of information associated with the building. Information includes but is not limited to: floor data, floor resource data, street address and location.
  */
 
 @interface PWBuilding : NSObject
@@ -45,7 +47,7 @@
 @property (copy, readonly) NSString *streetAddress;
 
 /**
- The location of the building in the lat/long coordinate space. This location is usually the center of the building. (read-only)
+ The location of the building in the lat/long coordinate space. This location is typically the center of the building. (read-only)
  */
 @property (readonly) CLLocationCoordinate2D location;
 
@@ -54,5 +56,11 @@
  */
 @property (readonly) NSArray *floors;
 
+/**
+ Finds a `PWBuildingFloor` object in the building that matches the provided floor identifier.
+ @param floorID The building floor identifier to use for the search.
+ @return The building's floor that matches the provided identifier, if one is found. If no matches are found, `nil` is returned.
+ */
+- (PWBuildingFloor*)floorForIdentifier:(PWBuildingFloorIdentifier)identifier;
 
 @end