+++ Framework/PWMapKit.framework/Versions/A/Headers/PWBuildingFloor.h
@@ -2,7 +2,7 @@
 //  PWMapFloor.h
 //  PWMapKit
 //
-//  Copyright (c) 2014 Phunware. All rights reserved.
+//  Copyright (c) 2015 Phunware. All rights reserved.
 //
 
 #import "PWMappingTypes.h"