+++ Framework/PWMapKit.framework/Versions/A/Headers/PWIndoorLocation+PWDirectionsWaypointProtocol.h
@@ -1,0 +1,15 @@
+//
+//  PWIndoorLocation+PWDirectionsWaypointProtocol.h
+//  PWMapKit
+//
+//  Created by Sam Odom on 2/17/15.
+//  Copyright (c) 2015 Phunware. All rights reserved.
+//
+
+#import <PWLocation/PWIndoorLocation.h>
+
+#import "PWDirectionsWaypointProtocol.h"
+
+@interface PWIndoorLocation (PWDirectionsWaypointProtocol) <PWDirectionsWaypointProtocol>
+
+@end