+++ Framework/PWMapKit.framework/Versions/A/Headers/PWRouteOverlayRenderer.h
@@ -2,7 +2,7 @@
 //  PWRouteOverlayRenderer.h
 //  PWMapKit
 //
-//  Copyright (c) 2014 Phunware. All rights reserved.
+//  Copyright (c) 2015 Phunware. All rights reserved.
 //
 
 #import <MapKit/MapKit.h>