+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 #PWMapKit Changelog
 
+##v2.4.0 (Friday, March 13th, 2015)
+
+* Added the ability to enable and disable annotation zoom levels. See the `annotationZoomLevelsEnabled` property on `PWMapView` for more information.
+* Added the ability for users to add and remove `PWBuildingAnnotation` objects to and from the map view. NOTE: When adding or removing annotations, only `annotationID` equality will be checked.
+* Fixed a bug where inaccessible routes were sometimes returned for accessible-only requests.
+* Fixed a bug where routing between two points on the same floor would sometimes return a route that spanned multiple floors.
+* Fixed a bug where setting `showsIndoorUserLocation` on `PWMapView to ‘NO’ would not hide the blue dot.
+* Fixed a bug where toggling route snapping tolerance would not resume route snapping when reenabled.
+* Added the ability to stop loading a building. See `stopLoadingBuilding` method on `PWMapView` for more information.
+* Fixed a bug where memory for the routing graph wasn't being released when the `PWMapView` instance was being deallocated. 
+* Fixed a bug where building assets would get ‘stuck’ while the user’s blue dot continued to move during a route across multiple floors (with a valid indoor location).
+* Added a flag to `PWBuildingAnnotationProtocol`, `occlusionEnabled`. This flag determines whether the annotation label for a given annotation will be occluded by other labels. If `occlusionEnabled` is set to `NO`, the annotation label will **always** be visible.
+* `PWMapView` now correctly calls the delegate callback ` -mapView:didChangeIndoorUserTrackingMode:`.
+* Added options that allow routing from any point of interest, user location or dropped pin annotation to any other point of interest, user location or dropped pin annotation (in other words, any type of route endpoint is now allowed on either end of the route). See `PWDirections` and `PWDirectionsWaypointProtocol` for more information.
+* Enhanced routing functionality now uses a synchronous directions request interface. The former asynchronous request/response model is still supported but will be deprecated. 
+* Renamed `PWBuildingOverlayDelegate` to `PWBuildingOverlayDelegateProtocol`.
+* Deprecated the old indoor location manager registration interface in favor of a more simply named method with accompanying unregister method. 
+* Removed the `-setCurrentFloor:` PWMapView property mutator declaration and removed the `readonly` attribute of the property. 
+
 ##v2.3.1 (Monday, March 2nd, 2015)
 * Fixed bug where when routing across multiple floors with a valid indoor location the build asset would get 'stuck' while the blue dot continued to move.
 * Updated `PWRouteOverlayRenderer` initializer to accept a building overlay as part of initialization: `- (instancetype)initWithRouteOverlay:(PWRouteOverlay *)overlay buildingOverlay:(PWBuildingOverlay *)buildingOverlay`